chore: Moving records bulk tests to the same file
