+++ argilla-server/tests/unit/api/handlers/v1/datasets/records/records_bulk/test_create_dataset_records_bulk.py
@@ -15,7 +15,14 @@
 from uuid import UUID
 
 import pytest
-from argilla_server.enums import DatasetStatus, QuestionType, ResponseStatus, SuggestionType
+from argilla_server.enums import (
+    DatasetStatus,
+    QuestionType,
+    ResponseStatus,
+    SuggestionType,
+    RecordStatus,
+    DatasetDistributionStrategy,
+)
 from argilla_server.models.database import Record, Response, Suggestion, User
 from httpx import AsyncClient
 from sqlalchemy import func, select
@@ -33,6 +40,7 @@ from tests.factories import (
     TextQuestionFactory,
     ChatFieldFactory,
     CustomFieldFactory,
+    AnnotatorFactory,
 )
 
 
@@ -647,3 +655,118 @@ class TestCreateDatasetRecordsBulk:
 
         assert response.status_code == 422
         assert (await db.execute(select(func.count(Record.id)))).scalar_one() == 0
+
+    async def test_create_dataset_records_bulk_updates_records_status(
+        self, db: AsyncSession, async_client: AsyncClient, owner: User, owner_auth_header: dict
+    ):
+        dataset = await DatasetFactory.create(
+            status=DatasetStatus.ready,
+            distribution={
+                "strategy": DatasetDistributionStrategy.overlap,
+                "min_submitted": 2,
+            },
+        )
+
+        user = await AnnotatorFactory.create(workspaces=[dataset.workspace])
+
+        await TextFieldFactory.create(name="prompt", dataset=dataset)
+        await TextFieldFactory.create(name="response", dataset=dataset)
+
+        await TextQuestionFactory.create(name="text-question", dataset=dataset)
+
+        response = await async_client.post(
+            self.url(dataset.id),
+            headers=owner_auth_header,
+            json={
+                "items": [
+                    {
+                        "fields": {
+                            "prompt": "Does exercise help reduce stress?",
+                            "response": "Exercise can definitely help reduce stress.",
+                        },
+                        "responses": [
+                            {
+                                "user_id": str(owner.id),
+                                "status": ResponseStatus.submitted,
+                                "values": {
+                                    "text-question": {
+                                        "value": "text question response",
+                                    },
+                                },
+                            },
+                            {
+                                "user_id": str(user.id),
+                                "status": ResponseStatus.submitted,
+                                "values": {
+                                    "text-question": {
+                                        "value": "text question response",
+                                    },
+                                },
+                            },
+                        ],
+                    },
+                    {
+                        "fields": {
+                            "prompt": "Does exercise help reduce stress?",
+                            "response": "Exercise can definitely help reduce stress.",
+                        },
+                        "responses": [
+                            {
+                                "user_id": str(owner.id),
+                                "status": ResponseStatus.submitted,
+                                "values": {
+                                    "text-question": {
+                                        "value": "text question response",
+                                    },
+                                },
+                            },
+                        ],
+                    },
+                    {
+                        "fields": {
+                            "prompt": "Does exercise help reduce stress?",
+                            "response": "Exercise can definitely help reduce stress.",
+                        },
+                        "responses": [
+                            {
+                                "user_id": str(owner.id),
+                                "status": ResponseStatus.draft,
+                                "values": {
+                                    "text-question": {
+                                        "value": "text question response",
+                                    },
+                                },
+                            },
+                            {
+                                "user_id": str(user.id),
+                                "status": ResponseStatus.draft,
+                                "values": {
+                                    "text-question": {
+                                        "value": "text question response",
+                                    },
+                                },
+                            },
+                        ],
+                    },
+                    {
+                        "fields": {
+                            "prompt": "Does exercise help reduce stress?",
+                            "response": "Exercise can definitely help reduce stress.",
+                        },
+                    },
+                ],
+            },
+        )
+
+        assert response.status_code == 201
+
+        response_items = response.json()["items"]
+        assert response_items[0]["status"] == RecordStatus.completed
+        assert response_items[1]["status"] == RecordStatus.pending
+        assert response_items[2]["status"] == RecordStatus.pending
+        assert response_items[3]["status"] == RecordStatus.pending
+
+        assert (await Record.get(db, UUID(response_items[0]["id"]))).status == RecordStatus.completed
+        assert (await Record.get(db, UUID(response_items[1]["id"]))).status == RecordStatus.pending
+        assert (await Record.get(db, UUID(response_items[2]["id"]))).status == RecordStatus.pending
+        assert (await Record.get(db, UUID(response_items[3]["id"]))).status == RecordStatus.pending
